# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "StudyForge Pro 智能学习平台 REST API。支持 AI 分析、间隔重复、流式对话、知识图谱等功能。\n\n所有认证端点需在 Header 携带 `Authorization: Bearer \u003ctoken\u003e`。",
        "title": "StudyForge Pro API",
        "contact": {
            "name": "StudyForge Pro",
            "url": "https://github.com/studyforge-pro"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "2.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/cards": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "分页获取知识卡片，支持按材料、难度、待复习状态、书签过滤",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "知识卡片"
                ],
                "summary": "卡片列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "每页条数（默认 20）",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "偏移量",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "按材料 ID 过滤",
                        "name": "material_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "按难度过滤（easy/medium/hard）",
                        "name": "difficulty",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "true 时只返回待复习卡片",
                        "name": "due",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "true 时只返回已书签卡片",
                        "name": "bookmarked",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "卡片列表",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/materials": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "分页获取当前用户的学习材料列表，支持标签过滤",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "材料管理"
                ],
                "summary": "材料列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "每页条数（默认 20，上限 200）",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "偏移量（默认 0）",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "按标签过滤",
                        "name": "tag",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "材料列表",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "创建一条文本或 URL 类型的学习材料",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "材料管理"
                ],
                "summary": "创建文本材料",
                "parameters": [
                    {
                        "description": "上传请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UploadRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "创建成功",
                        "schema": {
                            "$ref": "#/definitions/model.Material"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.Card": {
            "type": "object",
            "properties": {
                "concept": {
                    "description": "核心概念（一句话）",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "detail": {
                    "description": "详细解释",
                    "type": "string"
                },
                "difficulty": {
                    "description": "easy, medium, hard",
                    "type": "string"
                },
                "ease_factor": {
                    "description": "SM-2 难度因子（最低 1.3）",
                    "type": "number"
                },
                "formula": {
                    "description": "关键公式/代码（如有）",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "interval_days": {
                    "description": "当前复习间隔（天）",
                    "type": "integer"
                },
                "is_bookmarked": {
                    "description": "是否书签",
                    "type": "boolean"
                },
                "last_reviewed_at": {
                    "description": "上次复习时间",
                    "type": "string"
                },
                "material_id": {
                    "type": "string"
                },
                "memory_tip": {
                    "description": "记忆技巧",
                    "type": "string"
                },
                "next_review_at": {
                    "description": "下次复习时间（nil=新卡片，未复习过）",
                    "type": "string"
                },
                "review_count": {
                    "description": "累计复习次数",
                    "type": "integer"
                },
                "tags": {
                    "description": "逗号分隔的标签",
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "user_note": {
                    "description": "个人笔记",
                    "type": "string"
                }
            }
        },
...
        "model.Material": {
            "type": "object",
            "properties": {
                "analysis_data": {
                    "description": "JSON 格式的分析师输出",
                    "type": "string"
                },
                "analyzed_at": {
                    "type": "string"
                },
                "cards": {
                    "description": "关联",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Card"
                    }
                },
                "content": {
                    "description": "原文内容",
                    "type": "string"
                },
                "content_type": {
                    "description": "\"text\", \"pdf\", \"url\"",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "graph_data": {
                    "description": "JSON 格式的图谱数据",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_public": {
                    "description": "是否公开到市场",
                    "type": "boolean"
                },
                "quizzes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Quiz"
                    }
                },
                "share_code": {
                    "description": "分享码（公开时生成）",
                    "type": "string"
                },
                "source_url": {
                    "description": "来源 URL（如有）",
                    "type": "string"
                },
                "status": {
                    "description": "pending, analyzing, completed, failed",
                    "type": "string"
                },
                "tags": {
                    "description": "自定义标签，逗号分隔（如 \"编程,Go,并发\"）",
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "model.Quiz": {
            "type": "object",
            "properties": {
                "answer": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "difficulty": {
                    "description": "easy, medium, hard",
                    "type": "string"
                },
                "explanation": {
                    "description": "解析",
                    "type": "string"
                },
                "hint_1": {
                    "description": "第1级提示：概念方向",
                    "type": "string"
                },
                "hint_2": {
                    "description": "第2级提示：关键线索",
                    "type": "string"
                },
                "hint_3": {
                    "description": "第3级提示：接近答案",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "material_id": {
                    "type": "string"
                },
                "options": {
                    "description": "JSON 格式选项（选择题用）",
                    "type": "string"
                },
                "question": {
                    "type": "string"
                },
                "type": {
                    "description": "choice, fill, short_answer",
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "model.UploadRequest": {
            "type": "object",
            "required": [
                "content_type",
                "title"
            ],
            "properties": {
                "content": {
                    "type": "string"
                },
                "content_type": {
                    "type": "string",
                    "enum": [
                        "text",
                        "url"
                    ]
                },
                "source_url": {
                    "type": "string"
                },
                "tags": {
                    "description": "自定义标签，逗号分隔",
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT Bearer Token（格式: Bearer xxx）",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}